Create separate js files for update and clearall functionalities
