+++ src/style.css
@@ -50,6 +50,11 @@ ul li {
   font-size: 16px;
 }
 
+.add-form input:focus {
+  outline: none;
+  font-style: normal;
+}
+
 input[type=checkbox]:checked + input {
   text-decoration: line-through;
 }
@@ -57,12 +62,6 @@ input[type=checkbox]:checked + input {
 .description:read-only {
   text-decoration: line-through;
   font-style: italic !important;
-}
-
-
-.add-form input:focus {
-  outline: none;
-  font-style: normal;
 }
 
 .submit-button {
